requirements.txt
aioice==0.7.5
aiortc==1.2.1
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
av==8.0.3
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
cryptography==35.0.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.24
google-crc32c==1.2.0
greenlet==1.1.2
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
jsonschema==4.0.1
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.11.0
notebook==6.4.4
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pyarrow==5.0.0
pycparser==2.20
pydeck==0.7.0
pyee==8.2.2
Pygments==2.10.0
pylibsrtp==0.6.8
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
streamlit==0.89.0
streamlit-webrtc==0.29.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
torch==1.9.1
torchvision==0.10.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1